Finish a draft of the BlockStore Interface

+++ servers/src/main/java/tachyon/worker/block/BlockLock.java
@@ -1,19 +1,20 @@
 package tachyon.worker.block;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
  * A Lock to guard one block. There should be only one lock per block.
  */
 public class BlockLock {
-  static final AtomicInteger mBlockLockId = new AtomicInteger(0);
+  static final AtomicLong mBlockLockId = new AtomicLong(0);
 
   private final ReentrantLock mLock;
   /** The block Id this lock guards **/
   private final long mBlockId;
   /** The unique id of each lock **/
-  private final int mLockId;
+  private final long mLockId;
 
   public BlockLock(long blockId) {
     mBlockId = blockId;
@@ -25,7 +26,7 @@ public class BlockLock {
     return mBlockId;
   }
 
-  public int getLockId() {
+  public long getLockId() {
     return mLockId;
   }
